// src/app/models/user.ts
// Generated by https://quicktype.io

export interface UserModel {
    userId:    number;
    firstName: string;
    lastName:  string;
    userName:  string;
    bio:       string;
    imageId:   number;
    active:    boolean;
    image:     Image;
    events:    Event[];
}

export interface Event {
    eventId:     number;
    name:        string;
    description: string;
    imageId:     number;
    expired:     boolean;
    active:      boolean;
    image:       Image;
}

export interface Image {
    imageId: number;
    path:    string;
}